requirements.txt
pything @ git+https://github.com/thisgary/pything@main
discord.py
requests
